+++ tests/browser/stories.test.ts
@@ -32,6 +32,7 @@ const storySkipList = [
   'web-components-ds-accordion-item--default', // Redundant
   'web-components-ds-modal-dialog--default', // Doesn't show the open dialog
   'web-components-ds-modal-dialog--prevent-scroll-example', // Doesn't show the open dialog
+  'web-components-ds-autocomplete--async-items', // Redundant
 ];
 
 const storyUseAxeLegacyModeList = [
